🚚 move:"roomUtil 삭제 및 roomManger로 전부 변경"
