+++ game.server/src/services/take.damage.service.ts
@@ -1,4 +1,3 @@
-import { autoShieldBlock } from '../card/equip/card.auto_shield.effect';
 import { AnimationType, CardType, CharacterType } from '../generated/common/enums';
 import { playAnimationHandler } from '../handlers/play.animation.handler';
 import { cardManager } from '../managers/card.manager';
@@ -12,7 +11,8 @@ const takeDamageService = (room: Room, user: User, shooter: User, damage: number
 	const hasShield = user.character!.equips.includes(CardType.AUTO_SHIELD);
 	const isFroggy = user.character!.characterType === CharacterType.FROGGY;
 
-	const shieldRoll = hasShield && autoShieldBlock();
+	
+	const shieldRoll = hasShield &&  Math.random() < 0.25;
 	const froggyRoll = isFroggy && Math.random() < 0.25;
 
 	if (shieldRoll || froggyRoll) {
