+++ game.server/src/sockets/on.end.ts
@@ -2,12 +2,12 @@ import { handleError } from '../handlers/handleError.js';
 import { removeSocket } from '../managers/socket.manger.js';
 import { GameSocket } from '../type/game.socket.js';
 import { removeTokenUserDB } from '../services/prisma.service.js';
-import { deleteRoom, getRoom, removeUserFromRoom, saveRoom } from '../utils/room.utils.js';
 import { checkAndEndGameIfNeeded } from '../services/game.end.service.js';
 import { broadcastDataToRoom } from './notification.js';
 import { GamePacketType } from '../enums/gamePacketType.js';
 import { RoomStateType } from '../generated/common/enums.js';
 import { userUpdateNotificationPacketForm } from '../converter/packet.form.js';
+import roomManger from '../managers/room.manger.js';
 
 
 const onEnd = (socket: GameSocket) => async () => {
@@ -20,7 +20,7 @@ const onEnd = (socket: GameSocket) => async () => {
 		}
 
 		if (socket.roomId) {
-			const room = getRoom(Number(socket.roomId));
+			const room = roomManger.getRoom(Number(socket.roomId));
 			if (room) {
 				// 게임 중인 경우 플레이어를 죽은 상태로 처리
 				if (room.state === RoomStateType.INGAME) {
@@ -29,10 +29,7 @@ const onEnd = (socket: GameSocket) => async () => {
 												
 						// 플레이어 HP를 0으로 설정하여 죽은 상태로 처리
 						disconnectedUser.character.hp = 0;
-						
-						// 방 상태 저장
-						saveRoom(room);
-						
+										
 						// 모든 플레이어에게 사용자 상태 업데이트 알림 전송
 						const userUpdatePacket = userUpdateNotificationPacketForm(room.users);
 						await broadcastDataToRoom(room.users, userUpdatePacket, GamePacketType.userUpdateNotification);
@@ -43,11 +40,11 @@ const onEnd = (socket: GameSocket) => async () => {
 				}
 				
 				// 기존 로직: 방에서 사용자 제거
-				removeUserFromRoom(Number(socket.roomId), socket.userId!);
+				roomManger.removeUserFromRoom(Number(socket.roomId), socket.userId!);
 				
 				// 방에 사용자가 없으면 방 삭제
 				if (room.users.length <= 0) {
-					deleteRoom(Number(socket.roomId));
+					roomManger.deleteRoom(Number(socket.roomId));
 				}
 			}
 		}
