+++ game.server/src/useCase/create.room/create.room.usecase.ts
@@ -7,8 +7,6 @@ import { RoomStateType } from '../../generated/common/enums.js';
 import { User } from '../../models/user.model.js';
 import { createRoomDB, getUserByUserId } from '../../services/prisma.service.js';
 import { createRoomResponseForm } from '../../converter/packet.form';
-import roomManger from '../../managers/room.manger';
-import { saveRoom } from '../../utils/room.utils';
 
 const createRoomUseCase = async (
 	socket: GameSocket,
